Refactored + Updated Tests
