+++ src/App.tsx
@@ -1,13 +1,15 @@
 import { useEffect, useState } from "react";
+import { Route, Routes, useLocation } from "react-router-dom";
 import {
   AppShell,
   ColorScheme,
   ColorSchemeProvider,
   MantineProvider,
-  Paper,
   useMantineTheme,
 } from "@mantine/core";
-import { Route, Routes, useLocation } from "react-router-dom";
+import { useLocalStorage } from "@mantine/hooks";
+import { ApolloProvider } from "@apollo/client";
+import { QueryClientProvider } from "@tanstack/react-query";
 
 import { AppNavbar } from "./components/AppNavbar";
 import { AppFooter } from "./components/AppFooter";
@@ -17,12 +19,10 @@ import { Paginated } from "./pages/Paginated";
 import { InfiniteScroll } from "./pages/InfiniteScroll";
 import { BookDetails } from "./pages/BookDetails";
 
-import logo from "./assets/logo.png";
-import { ApolloProvider } from "@apollo/client";
 import { apolloClient } from "./libraries/apollo";
-import { QueryClientProvider } from "@tanstack/react-query";
 import { queryClient } from "./libraries/react-query";
-import { useHotkeys, useLocalStorage } from "@mantine/hooks";
+
+import logo from "./assets/logo.png";
 
 export default function App() {
   const theme = useMantineTheme();
@@ -42,8 +42,6 @@ export default function App() {
 
   const toggleColorScheme = (value?: ColorScheme) =>
     setColorScheme(value || (colorScheme === "dark" ? "light" : "dark"));
-
-  useHotkeys([["mod+J", () => toggleColorScheme()]]);
 
   return (
     <ApolloProvider client={apolloClient}>
