+++ src/setupTests.ts
@@ -2,4 +2,6 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom';
+import "@testing-library/jest-dom";
+
+global.ResizeObserver = require("resize-observer-polyfill");
